requirements.txt
streamlit
faiss-cpu
sentence-transformers
google-generativeai
numpy
python-docx